refactor: [0000004] - extract the different coulour creation/modification/generation functions from ColourDetailsActivity to a wider-access utill-file top-level-funcs.
